requirements.txt
Django>=3.0,<4.0
mysqlclient==1.4.6
django-crispy-forms==1.9.2
django-tempus-dominus==5.1.2.13
django-bootstrap-modal-forms==2.0.0